# api/docs/swagger.yaml
definitions:
  models.Book:
    properties:
      author_name:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      page_number:
        type: integer
      updated_at:
        type: string
    type: object
  models.BooksResponse:
    properties:
      books:
        items:
          $ref: '#/definitions/models.Book'
        type: array
      count:
        type: integer
    type: object
  models.CreateBook:
    properties:
      author_name:
        type: string
      name:
        type: string
      page_number:
        type: integer
    type: object
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
  models.UpdateBook:
    properties:
      author_name:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.UpdatePageNumber:
    properties:
      id:
        type: string
      page_number:
        type: integer
    type: object
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
  /book:
    post:
      consumes:
      - application/json
      description: create a new book
      parameters:
      - description: book
        in: body
        name: book
        schema:
          $ref: '#/definitions/models.CreateBook'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Creates a new book
      tags:
      - book
  /book/{id}:
    delete:
      consumes:
      - application/json
      description: delete book
      parameters:
      - description: book_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete book
      tags:
      - book
    get:
      consumes:
      - application/json
      description: get book by id
      parameters:
      - description: book_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get book by id
      tags:
      - book
    patch:
      consumes:
      - application/json
      description: update book page number
      parameters:
      - description: book_id
        in: path
        name: id
        required: true
        type: string
      - description: book
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePageNumber'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update book page number
      tags:
      - book
    put:
      consumes:
      - application/json
      description: update book
      parameters:
      - description: book_id
        in: path
        name: id
        required: true
        type: string
      - description: book
        in: body
        name: book
        schema:
          $ref: '#/definitions/models.UpdateBook'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update book
      tags:
      - book
  /books:
    get:
      consumes:
      - application/json
      description: get book list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BooksResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get book list
      tags:
      - book
swagger: "2.0"